fix compiler error on 64bits machines:
CElmPopup.cc: In function 'void elm::_item_selected_cb(void*, Evas_Object*, void*)':
CElmPopup.cc:245:30: error: cast from 'void*' to 'int' loses precision [-fpermissive]

CElmPopup.cc: In member function 'v8::Handle<v8::Value> elm::CElmPopup::message(const v8::Arguments&)':
CElmPopup.cc:320:120: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]
CElmPopup.cc:323:113: warning: cast to pointer from integer of different size [-Wint-to-pointer-cast]



git-svn-id: http://svn.enlightenment.org/svn/e/trunk@73782 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33

+++ PROTO/elev8/src/modules/elm/CElmPopup.cc
@@ -242,7 +242,7 @@ static void _item_selected_cb(void *data, Evas_Object *popup, void *)
 {
    CElmPopup *self = static_cast<CElmPopup *>(evas_object_data_get(popup, "this"));
    if (self)
-     self->ItemSelected((int)data);
+     self->ItemSelected((int)(long)data);
 }
 
 Handle<Value> CElmPopup::message(const Arguments &args)
@@ -317,10 +317,10 @@ Handle<Value> CElmPopup::message(const Arguments &args)
                   Evas_Object *eo_icon = (!icon.IsEmpty() && icon->IsObject()) ?
                      GetEvasObjectFromJavascript(Realise(icon->ToObject(), jsObject)) : NULL;
 
-                   elm_popup_item_append(eo, *String::Utf8Value(text->ToString()), eo_icon, _item_selected_cb, (void *)i);
+                   elm_popup_item_append(eo, *String::Utf8Value(text->ToString()), eo_icon, _item_selected_cb, (void *)(long)i);
                }
              else
-               elm_popup_item_append(eo, *String::Utf8Value(item->ToString()), NULL, _item_selected_cb, (void *)i);
+               elm_popup_item_append(eo, *String::Utf8Value(item->ToString()), NULL, _item_selected_cb, (void *)(long)i);
           }
      }
